renamed db folders; renamed sqllite connection variable.
